backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
gitpython>=3.1.42
pydantic>=2.7.1
